systemd-analyze wants systemd finished to boot
When called from mailboot.service, the systemd-analyze output was missing
because systemd-analyze print the following line on stderr (not in mail):
Bootup is not yet finished. Please try again later.

To workaround this, we will test systemd-analyze during a configurable time.
If systemd-analyze is not ready at the time left, the information will be
ignored.

systemd unit type is now Tyep=idle, to:
- don't block (particulary when waiting systemd-analyze)
- we don't need to be executed fastly (we can be more fair than Type=simple).